better handling of pulse guide failure (Fixes Issue 262)   Mount::Guide now returns one of three possible values: move succeeded, move failed, and     move failed and guiding must stop. The third value is used to stop guiding when     slewing is detected.   Workaround a Gemini2 bug that was causing PHD2 to hang when the guide pulse failed to     complete by calling AbortSlew when the condition is detected.   Also includes a partial fix for Issue 264 by including the AO tilt position in the Event     Server GuideStep event.   Also, center the connect equipment dialog in the main window whenever it is displayed,     not just the first time.

+++ graph.h
@@ -58,8 +58,8 @@ struct S_HISTORY
     double dy;
     double ra;
     double dec;
-    double raDur;
-    double decDur;
+    int raDur;
+    int decDur;
     double starSNR;
     double starMass;
     S_HISTORY() { }
